% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{COGU_sampled_spmapper}
\alias{COGU_sampled_spmapper}
\title{Common Guillemot sampled parameters}
\format{
A data frame with 60000 rows and 17 variables:
\describe{
\item{\code{TB_nest}}{double Time-activity budget: nest}
\item{\code{TB_flying}}{double Time-activity budget: flight}
\item{\code{TB_resting_sea}}{double Time-activity budget: rest-at-sea}
\item{\code{TB_foraging}}{double Time-activity budget: foraging}
\item{\code{BS_length}}{integer Number of days chick-rearing}
\item{\code{ER_foraging}}{double Energetic cost: foraging}
\item{\code{ER_flying}}{double Energetic cost: flight}
\item{\code{ER_resting_sea}}{double Energetic cost: rest at sea}
\item{\code{ER_nest}}{double Energetic cost: nest}
\item{\code{ER_warming_food}}{double Energetic cost: warming food}
\item{\code{chicks_n}}{double Number of chicks}
\item{\code{chick_ER}}{double Daily chick energetic requirement}
\item{\code{assim_effic}}{double Digestive assimilation efficiency}
\item{\code{init_mass}}{double Initial adult body mass}
\item{\code{mass_change}}{double Breeding season change in adult body mass}
\item{\code{tissue_en_dens}}{double Adult seabird tissue energy density}
\item{\code{energy_prey_conv}}{double Prey energy density}
}
}
\source{
Randomly sampled deviates of parameters from 'Seabird_energetics_pars_spmapper'. See methods for details.
}
\usage{
COGU_sampled_spmapper
}
\description{
This dataset contains random deviate samples of Guillemot energetics & behavioural parameters, based on 'Seabird_energetics_pars_spmapper'. Each row represent a set of parameter values used by the prey consumption model. See methods for details.
}
\keyword{datasets}
