% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fame_population_spmapper}
\alias{fame_population_spmapper}
\title{Fame population}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{\code{spp}}{character COLUMN_DESCRIPTION}
\item{\code{estimate}}{character COLUMN_DESCRIPTION}
\item{\code{value}}{integer COLUMN_DESCRIPTION}
}
}
\source{
NEED EXPL
}
\usage{
fame_population_spmapper
}
\description{
A dataset containing information on:
}
\keyword{datasets}
