% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FAME_UD_standardised_COGU}
\alias{FAME_UD_standardised_COGU}
\title{FAME-UD-standardised-COGU-spmapper}
\format{
A list with two elements:
\describe{
\item{matrix}{A numeric matrix of UD values with 462,392 rows and 1 column (with bird species code).
\item{metadata}{A list of 7 elements describing spatial extent and resolution:
\describe{
\item{n_rows}{Number of raster rows}
\item{n_cols}{Number of raster columns}
\item{x_min}{Minimum X coordinate}
\item{x_max}{Maximum X coordinate}
\item{y_min}{Minimum Y coordinate}
\item{y_max}{Maximum Y coordinate}
\item{crs}{Coordinate reference system as WKT or PROJ string}
}
}
}
}
\source{
Wakefield, E.D., Owen, E., Baer, J., Carroll, M.J., Daunt, F., Dodd, S.G., Green, J.A., Guilford, T., Mavor, R.A., Miller, P.I. and Newell, M.A., 2017. Breeding density, fine‐scale tracking, and large‐scale modeling reveal the regional distribution of four seabird species. Ecological Applications, 27(7), pp.2074-2091.
}
\usage{
FAME_UD_standardised_COGU
}
\description{
Uk-Ireland utilisation distribution (UD) for Common Guillemot (Future of the Atlantic Marine Environment FAME study period; 2010-2014). Via RSPB; for methods, see Wakefield et al. (2027), Ecological Applications.
}
\details{
The object contains a matrix of UD values and associated spatial metadata, to build a raster
}
\keyword{datasets}
