% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FAME_UD_standardised_RAZO}
\alias{FAME_UD_standardised_RAZO}
\title{FAME-UD-standardised-RAZO-spmapper}
\format{
A list with two elements:
\describe{
\item{matrix}{A numeric matrix of UD values with 462,392 rows and 1 column (with bird species code).
\item{metadata}{A list of 7 elements describing spatial extent and resolution:
\describe{
\item{n_rows}{Number of raster rows}
\item{n_cols}{Number of raster columns}
\item{x_min}{Minimum X coordinate}
\item{x_max}{Maximum X coordinate}
\item{y_min}{Minimum Y coordinate}
\item{y_max}{Maximum Y coordinate}
\item{crs}{Coordinate reference system as WKT or PROJ string}
}
}
}
}
\source{
NEED EXPL
}
\usage{
FAME_UD_standardised_RAZO
}
\description{
A dataset containing information on:
}
\details{
The object contains a matrix of UD values and associated spatial metadata, to build a raster
}
\keyword{datasets}
