% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_fpudoverlap.R
\name{fpudoverlap}
\alias{fpudoverlap}
\title{Calculate proportion of UD that lies within polygon(s)}
\usage{
fpudoverlap(fppolys, udmap)
}
\arguments{
\item{fppolys}{Footprint(s): a simple feature collection}

\item{udmap}{A utilisation distribution: NOTE: values must sum to approximately one, but not currently checked}
}
\value{
Proportion of UD overlapping a polygon: a single numeric value
}
\description{
Calculate proportion of UD that lies within polygon(s)
}
