% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RAZO_sampled_spmapper}
\alias{RAZO_sampled_spmapper}
\title{Razorbill sampled parameters}
\format{
A data frame with 60000 rows and 17 variables:
\describe{
\item{\code{TB_nest}}{double COLUMN_DESCRIPTION}
\item{\code{TB_flying}}{double COLUMN_DESCRIPTION}
\item{\code{TB_resting_sea}}{double COLUMN_DESCRIPTION}
\item{\code{TB_foraging}}{double COLUMN_DESCRIPTION}
\item{\code{BS_length}}{integer COLUMN_DESCRIPTION}
\item{\code{ER_foraging}}{double COLUMN_DESCRIPTION}
\item{\code{ER_flying}}{double COLUMN_DESCRIPTION}
\item{\code{ER_resting_sea}}{double COLUMN_DESCRIPTION}
\item{\code{ER_nest}}{double COLUMN_DESCRIPTION}
\item{\code{ER_warming_food}}{double COLUMN_DESCRIPTION}
\item{\code{chicks_n}}{double COLUMN_DESCRIPTION}
\item{\code{chick_ER}}{double COLUMN_DESCRIPTION}
\item{\code{assim_effic}}{double COLUMN_DESCRIPTION}
\item{\code{init_mass}}{double COLUMN_DESCRIPTION}
\item{\code{mass_change}}{double COLUMN_DESCRIPTION}
\item{\code{tissue_en_dens}}{double COLUMN_DESCRIPTION}
\item{\code{energy_prey_conv}}{double COLUMN_DESCRIPTION}
}
}
\source{
NEED EXPL
}
\usage{
RAZO_sampled_spmapper
}
\description{
This dataset contains information about,...
}
\keyword{datasets}
