% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_spmapplot.R
\name{spmapplot}
\alias{spmapplot}
\title{Plot prey consumption maps, with user polygons overlaid (optional)}
\usage{
spmapplot(rast, polyg = NULL)
}
\arguments{
\item{rast}{A spatRaster raster object to be plotted (i.e. prey consumption map)}

\item{polyg}{An optional sf object (i.e. user polygons) to overlay on the raster plot (default: \code{NULL}).}
}
\value{
Prey consumption map with Western European country polygons (via rnaturalearth), with (optional) user input polygons overlaid: ggplot
}
\description{
Plot prey consumption maps, with user polygons overlaid (optional)
}
