% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fame_population_spmapper}
\alias{fame_population_spmapper}
\title{Fame population}
\format{
A data frame with 3 rows and 2 variables:
\describe{
\item{\code{spp}}{character Species}
\item{\code{value}}{integer Breeding population size estimate}
}
}
\source{
See methods for details.
}
\usage{
fame_population_spmapper
}
\description{
A dataset containing UK seabird population size estimates for breeding population size of focal seabird species. A full national census (Seabird 2000) is available for the period 1998-2002, whereas colony counts for a subset of colonies are available for the period 2010-2014 via the Seabird Monitoring Programme (SMP). We used both datasets to derive an estimate of national-level abundance for the period 2010-2014. See methods for further details.
}
\keyword{datasets}
