% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Seabird_energetics_pars_spmapper}
\alias{Seabird_energetics_pars_spmapper}
\title{Seabirds Energetics Parameters}
\format{
A data frame with the following 4 columns:
\describe{
\item{parameter}{Character: ...}
\item{Guillemot}{Numeric: ...}
\item{Razorbill}{Numeric: ...}
\item{Kittiwake}{Numeric: ...}
}
}
\source{
NEED EXPL
}
\usage{
Seabird_energetics_pars_spmapper
}
\description{
This dataset provides information about :
}
\keyword{datasets}
