% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_spmapper.R
\name{spmapper}
\alias{spmapper}
\title{TO BE ADDED - TITLE OF MAIN SPMAPPER FUNCTION}
\usage{
spmapper(spname, fppolys, popscen, tooldir)
}
\arguments{
\item{spname}{Species name: a character string. Must be "Guillemot", "Kittiwake" or "Razorbill"}

\item{fppolys}{Footprint(s): a simple feature collection}

\item{popscen}{Population scenario: a character string. Must be "baseline", "low" or "high"}

\item{tooldir}{Path to directory containing tool: a character string.}
}
\value{
A list, with elements: "spname" (species name), "popscen" (population scenario"), "energypars_base_est" (energetics parameters - estimates), "energypars_base_sim" (energetics parameters - simulations), "udmap" (utilisation distribution, a raster), "popsize" (population size - BREEDING PAIRS \strong{TBC}, "fp_ud_overlap" (proportion of UD within a footprint), "nbirds_in_fp" (number of birds in footprints), "onebird_prey_est" (prey intake from one bird - estimate), "onebird_prey_sim" (prey intake from one bird - simulations), "allbirds_prey_est" (prey intake from all birds in footprints - estimate), "allbirds_prey_sim" (prey intake from all birds in footprints - simulations)
}
\description{
TO BE ADDED - OVERALL DESCRIPTION OF WHAT SPMAPPER DOES
}
