% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_spmapper.R
\name{spmapper}
\alias{spmapper}
\title{spmapper core function. Generate prey consumption maps and perform extractions within user-supplied polygons.}
\usage{
spmapper(spname, fppolys, popscen, tooldir)
}
\arguments{
\item{spname}{Species name: a character string. Must be "Guillemot", "Kittiwake" or "Razorbill"}

\item{fppolys}{Footprint(s): a simple feature collection}

\item{popscen}{Population scenario: a character string. Must be "baseline", "low" or "high"}

\item{tooldir}{Path to directory containing tool: a character string.}
}
\value{
A list, with elements: "spname" (species name), "popscen" (population scenario"), "energypars_base_est" (energetics parameters - estimates), "energypars_base_sim" (energetics parameters - simulations), "udmap" (utilisation distribution, a raster), "popsize" (population size - number of individual breeding adults), "fp_ud_overlap" (proportion of UD within polygons/footprints), "onebird_prey_kg_est" (chick-rearing prey intake mass from one bird - estimate), "onebird_prey_kg_sim" (chick-rearing prey intake mass from one bird - simulations), "allbirds_prey_est" (prey intake mass from all birds - estimate),  "allbirds_prey_sim" (prey intake from all birds - simulations), "allbirds_prey_fp_est" (prey mass in footprints - estimate),  "allbirds_prey_sim" (prey mass in footprints - simulations), "prey_cons_map" (prey consumption map - a raster)
}
\description{
Produces prey consumption maps and extracts of prey mass values within user-specified spatial polygons. Returns list of various inputs, intermediate quantities, prey consumption maps and extraction totals.
}
